Add theming to web pages
