+++ package.json
@@ -22,6 +22,7 @@
     "cookie-parser": "1.4.6",
     "debug": "4.3.4",
     "express": "4.18.2",
+    "govuk-frontend": "^4.3.1",
     "morgan": "1.10.0",
     "node-plop": "0.26.3",
     "nunjucks": "3.2.3",
